utils.js & now the Home page can mirror  the changes!
