+++ src/js/elements.js
@@ -1,6 +1,6 @@
 const modalInner = document.querySelector('.modal-inner');
 const modalOuter = document.querySelector('.modal-outer');
-const editableTexts = Array.from(document.querySelectorAll('.edit-text'));
+const editableElements = Array.from(document.querySelectorAll('.edit-text'));
 const editButton = document.querySelector('.edit-button');
 const loginButtons = Array.from(document.querySelectorAll('.login'));
-export { modalInner, modalOuter, editableTexts, editButton, loginButtons };
+export { modalInner, modalOuter, editableElements, editButton, loginButtons };
