+++ src/js/loguin.js
@@ -1,55 +1,24 @@
 import {
   modalInner,
   modalOuter,
-  editableTexts,
+  editableElements,
   editButton,
   loginButtons,
 } from './elements';
+import { mirrorToLocalStorage } from './utils';
 
 let loginForm;
-let reducedEditables;
-function editableTextsReducer(arr) {
-  const reduced = arr.reduce((textObj, currentEl) => {
-    textObj[currentEl.id] = currentEl.innerText;
-    return textObj;
-  }, {});
-  return reduced;
-}
-
-function handleEditText(e) {
-  const el = e.target;
-}
-function setEventsListeners() {
-  editableTexts.forEach((el) => {
-    el.addEventListener('input', handleEditText);
-  });
-}
-
-function mirrorToLocalStorage() {
-  reducedEditables = editableTextsReducer(editableTexts);
-  const { title } = document;
-  console.log(title);
-  switch (title) {
-    case 'Home':
-      localStorage.setItem('Home', JSON.stringify(reducedEditables));
-      break;
-    case 'Experience':
-      localStorage.setItem('Experience', JSON.stringify(reducedEditables));
-      break;
-    default:
-      break;
-  }
-}
 
 function handleSubmit(e) {
   e.preventDefault();
   modalOuter.classList.remove('open');
   editButton.hidden = false;
-  editableTexts.forEach((el) => {
+  editableElements.forEach((el) => {
     el.contentEditable = true;
   });
-  mirrorToLocalStorage();
-  setEventsListeners();
+  setInterval(() => {
+    mirrorToLocalStorage();
+  }, 10000);
 }
 function createForm() {
   modalOuter.classList.add('open');
