+++ src/js/main.js
@@ -2,6 +2,7 @@ import dropdown from 'bootstrap/js/dist/dropdown.js';
 // import popover from 'bootstrap/js/dist/popover.js';
 import { Slider1 } from './slider.js';
 import { manageLogin } from './loguin.js';
+import { restoreFromLStorage } from './utils';
 /*
 const popers = document.querySelectorAll('[data-bs-toggle="popover"]');
 Array.from(popers).forEach((pop)=>new bootstrap.Popover(pop));
@@ -40,3 +41,4 @@ window.addEventListener('resize', handleResize);
 
 const myHomeSlider = new Slider1(document.querySelector('.home-slider'));
 manageLogin();
+restoreFromLStorage();
